requirements.txt
django
djangorestframework
xlsxwriter